Extract stable-placement search into placement_search.rs
Move the promote-vs-shadow local search (problem/state types, move
search, packing-entry builders, cost model) and its tests out of
memory_plan.rs, which now holds the plan data model and program-level
assembly only. No behavior change.
